requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
pydantic-settings==2.*
aiosqlite==0.20.*
httpx==0.28.*
PyJWT[crypto]==2.10.*
cryptography==44.*
pyyaml==6.*
pytest==8.*
pytest-asyncio==0.24.*
httpx  # also used as test client via httpx.AsyncClient
